Java updates for beta
I was hoping for a more refined set of changes, but it became sort of a
grab bag. Some fixes are things that didn't display correctly. Others
are Java changes that happened leading up to beta. And yet others are
changes just to not lead the user in a painful direction.