Prevent lists and h3 from being distracting (#271)
* Lists should match paragraphs in feel

They were previously too dark and stood out. They may now be slightly
too hard to read, because paragraphs are already bordering on
readability issues and lists are 1px smaller. Good enough for now
though.

* h3 should stand out less than h2

The black made it stand out a lot when glancing/scrolling down the page.
It now is something between h2 and h4 in standing out. It could probably
be tweaked further, but it is no longer a glaring issue. This was quite
noticeable in the java tutorial.