Update auth docs for secure ruby connections.
The existing ruby documentation is out-of-date (uses Credentials instead of ChannelCredentials) and doesn't document the arguments for secure connections.  Unfortunately, there is no auto-generated ruby documentation that I know of that actually pulls together the constructor args and the call options, so once I figured it out I wanted to add it here so that this documentation is useful when creating secure RPCs.